% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsis.R
\name{reweight_stanfit_draws}
\alias{reweight_stanfit_draws}
\title{Update post-warmup draws of a stanfit based on a vector of weights}
\usage{
reweight_stanfit_draws(stanfit, draw_weights)
}
\arguments{
\item{stanfit}{a stanfit}

\item{weights}{a vector of weights}
}
\description{
Update post-warmup draws of a stanfit based on a vector of weights
}
