% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsis.R
\name{upsis.brmsfit}
\alias{upsis.brmsfit}
\title{Update a brms model using PSIS}
\usage{
\method{upsis}{brmsfit}(model, data_add = NULL, data_remove = NULL)
}
\arguments{
\item{model}{a brmsfit object}

\item{data_add}{additional data collected after the model was fit}

\item{data_remove}{data included in model fitting whose influence to remove}
}
\description{
Update a brms model using PSIS
}
