% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsis.R
\name{upsis}
\alias{upsis}
\title{Update a model using PSIS}
\usage{
upsis(model, data_add = NULL, data_remove = NULL)
}
\arguments{
\item{model}{a model object (currently only brmsfit supported)}

\item{data_add}{additional data collected after the model was fit}

\item{data_remove}{data included in model fitting whose influence to remove}
}
\description{
Update a model using PSIS
}
